Move data structure to its own file (#4555)
* Move data structure to its own class

* Fix pre-commit failures

* Create data_structure submodule in core
