+++ process/core/data_structure/__init__.py
@@ -1,0 +1,1 @@
+"""Contains classes related to the PROCESS data structure."""
