+++ process/core/solver/solver.py
@@ -17,8 +17,8 @@ from pyvmcon import (
 from scipy.optimize import fsolve
 
 from process.core import constants, process_output
+from process.core.data_structure.base import DataStructure
 from process.core.exceptions import ProcessValueError
-from process.core.model import DataStructure
 from process.core.solver.evaluators import Evaluators
 from process.data_structure.numerics import SolverOutputCondition
