+++ process/main.py
@@ -39,13 +39,14 @@ import click
 
 import process  # noqa: F401
 from process.core import constants, init
+from process.core.data_structure.base import DataStructure
 from process.core.io import obsolete_vars as ov
 from process.core.io.cli_tools import LazyGroup, help_opt, indat_opt
 from process.core.io.mfile import MFile
 from process.core.io.plot import plot_sankey_plotly, plot_summary
 from process.core.io.vary_run import RunProcessConfig, vary_iteration_variables
 from process.core.log import logging_model_handler, show_errors
-from process.core.model import DataStructure, Model
+from process.core.model import Model
 from process.core.process_output import OutputFileManager, oheadr
 from process.core.scan import Scan
 from process.data_structure.blanket_variables import BlktModelTypes
