+++ process/models/stellarator/coils/quench.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 
-from process.core.model import DataStructure
+from process.core.data_structure.base import DataStructure
 
 
 def calculate_quench_protection(coilcurrent, data: DataStructure):
