+++ process/models/stellarator/divertor.py
@@ -4,7 +4,7 @@ import numpy as np
 
 from process.core import constants
 from process.core import process_output as po
-from process.core.model import DataStructure
+from process.core.data_structure.base import DataStructure
 
 
 def st_div(stellarator, f_output: bool, data: DataStructure):
