+++ process/models/stellarator/heating.py
@@ -3,8 +3,8 @@
 import logging
 
 from process.core import process_output as po
+from process.core.data_structure.base import DataStructure
 from process.core.exceptions import ProcessValueError
-from process.core.model import DataStructure
 from process.data_structure.physics_variables import PlasmaIgnitionModel
 
 logger = logging.getLogger(__name__)
