+++ process/models/tfcoil/superconducting.py
@@ -10,8 +10,8 @@ from scipy import optimize
 
 from process.core import constants
 from process.core import process_output as po
+from process.core.data_structure.base import DataStructure
 from process.core.exceptions import ProcessValueError
-from process.core.model import DataStructure
 from process.data_structure.superconducting_tf_coil_variables import TFWPIntegerTurnType
 from process.models import superconductors
 from process.models.superconductors import (
