+++ tests/unit/core/test_data_structure.py
@@ -2,7 +2,7 @@ from dataclasses import fields
 
 import pytest
 
-from process.core.model import DataStructure
+from process.core.data_structure.base import DataStructure
 
 DATA_STRUCTURES = {f.name: f.type() for f in fields(DataStructure)}
